Updated collision detection system:
ROUNDED COLLISION CHECK

Normally collision is checked based on the object's rectangular dimensions, which can be a bit clunky.

For objects such as a player, you can set the following attribute:

obj.rounded_collision_check_ = True #normally this is False

This creates a circular area inside the object for collision checking.